requirements.txt
alembic==1.11.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
SQLAlchemy==1.3.17
psycopg2-binary==2.8.5
requests==2.24.0
python-dateutil==2.8.1
pytz==2020.1
Jinja2==2.11.2
Werkzeug==1.0.1
Pillow==7.2.0
sqlalchemy-media==0.17.4
psycopg2-binary
python-dotenv==1.0.0